update components + add some systems

+++ core/src/main/java/com/populaire/projetguerrefroide/component/Population.java
@@ -3,5 +3,5 @@ package com.populaire.projetguerrefroide.component;
 import com.github.elebras1.flecs.annotation.Component;
 
 @Component
-public record Population(int index, long typeId, long provinceId, long amount, long employment, float consciousness, float militancy, float literacy, float savings) {
+public record Population(int index, long typeId, long provinceId, int amount, int employment, float consciousness, float militancy, float literacy, float savings) {
 }
